feat: update script
