+++ scripts/publish-beta-extension.ts
@@ -2,38 +2,11 @@
  * Scripts to check unpublished version and run beta publish
  */
 import * as path from 'path';
-import * as fs from 'fs-extra';
 import { spawnSync } from 'child_process';
 import uploadExtesions from './upload-extensions';
-import { getPublishedPackages } from './published-info';
 import { IExtensionInfo, getExtensionInfos } from './getExtensionInfos';
 import extensionDepsInstall from './fn/extension-deps-install';
-
-function updateBetaDependencies(extension: string, directory: string) {
-  try {
-    const publishedPackages: string[] = getPublishedPackages();
-
-    if (fs.existsSync(directory)) {
-      const packageFile = path.join(directory, 'package.json');
-      const packageData = fs.readJsonSync(packageFile);
-
-      publishedPackages.forEach((publishedPackage: string) => {
-        const info = publishedPackage.split(':');
-        const name = info[0];
-        const version = info[1];
-
-        if (packageData.dependencies && packageData.dependencies[name]) {
-          packageData.dependencies[name] = version;
-        } else if (packageData.devDependencies && packageData.devDependencies[name]) {
-          packageData.devDependencies[name] = version;
-        }
-      });
-      fs.writeFileSync(packageFile, JSON.stringify(packageData, null, 2));
-    }
-  } catch (e) {
-    console.log(`[ERROR] ${extension} update beta package dependencies failed.`, e);
-  }
-}
+import updateBetaDependencies from './fn/updateExtensionDependencies';
 
 function publish(extension: string, directory: string, version: string): void {
   // vsce package
