+++ scripts/publish-extension.ts
@@ -1,12 +1,20 @@
 /**
  * Scripts to check unpublished version and run publish
  */
+import * as path from 'path';
 import { spawnSync } from 'child_process';
 import { getLatestVersion } from 'ice-npm-utils';
 import uploadExtesions from './upload-extensions';
 import { getPublishedPackages } from './published-info';
 import { IExtensionInfo, getExtensionInfos } from './getExtensionInfos';
 import extensionDepsInstall from './fn/extension-deps-install';
+import updateBetaDependencies from './fn/updateExtensionDependencies';
+
+// Wait for npm module publish and sync.
+function sleep(time) {
+  console.log(`start sleep: ${time} ms`);
+  return new Promise((resolve) => setTimeout(resolve, time));
+}
 
 // Check published packages can be installed.
 function checkPackagePublished() {
@@ -64,38 +72,53 @@ function publish(extension: string, directory: string, version: string): void {
   });
 }
 
-checkPackagePublished()
-  .then(() => {
-    // Entry
+async function start() {
+  try {
+    await checkPackagePublished();
+    await sleep(50000);
     console.log('[PUBLISH] Start:');
-    getExtensionInfos().then((extensionInfos: IExtensionInfo[]) => {
-      // npm install
-      extensionDepsInstall();
 
-      // Publish
-      let publishedCount = 0;
-      const extensions = [];
-      const publishedExtensions = [];
-      for (let i = 0; i < extensionInfos.length; i++) {
-        const { name, directory, localVersion, shouldPublish } = extensionInfos[i];
-        // Production publish should zip all extensions.
-        // Pack extension first.
-        extensions.push(`${name}:${localVersion}`);
-        packExtension(name, directory, localVersion);
+    const extensionInfos: IExtensionInfo[] = await getExtensionInfos();
+    const shouldPublishExtensions: IExtensionInfo[] = [];
 
-        if (shouldPublish) {
-          publishedCount++;
-          console.log(`--- ${name}@${localVersion} ---`);
-          publish(name, directory, localVersion);
-          publishedExtensions.push(`${name}:${localVersion}`);
-        }
+    for (let i = 0; i < extensionInfos.length; i++) {
+      const { name, directory, shouldPublish } = extensionInfos[i];
+      if (shouldPublish) {
+        // Update extension package json
+        updateBetaDependencies(name, directory);
+        // Update inside web project package json
+        updateBetaDependencies(name, path.join(directory, 'web'));
+  
+        shouldPublishExtensions.push(extensionInfos[i]);
       }
-      uploadExtesions(extensions, true);
-      console.log(`[PUBLISH EXTENSION PRODUCTION] Complete (count=${publishedCount}):`);
-      console.log(`${publishedExtensions.join('\n')}`);
-    });
-  })
-  .catch((e) => {
+    }
+
+    // npm install
+    extensionDepsInstall();
+
+    // Publish
+    let publishedCount = 0;
+    const extensions = [];
+    const publishedExtensions = [];
+    for (let i = 0; i < shouldPublishExtensions.length; i++) {
+      const { name, directory, localVersion } = extensionInfos[i];
+      // Production publish should zip all extensions.
+      // Pack extension first.
+      extensions.push(`${name}:${localVersion}`);
+      packExtension(name, directory, localVersion);
+        publishedCount++;
+        console.log(`--- ${name}@${localVersion} ---`);
+        publish(name, directory, localVersion);
+        publishedExtensions.push(`${name}:${localVersion}`);
+    }
+    uploadExtesions(extensions, true);
+    console.log(`[PUBLISH EXTENSION PRODUCTION] Complete (count=${publishedCount}):`);
+    console.log(`${publishedExtensions.join('\n')}`);
+
+  } catch (e) {
     console.error(e);
     process.exit(1);
-  });
+  }
+}
+
+start();
